% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_raster.R
\name{aggregate_raster}
\alias{aggregate_raster}
\title{Downscale raster resolution}
\usage{
aggregate_raster(rast, outfile, aggregate_factor, fun = sum,
  return_rast = FALSE)
}
\arguments{
\item{rast}{Raster* object or file path to a raster file.}

\item{outfile}{Character. Output raster file path. If not provided, object 
will be returned to R. Directory will be created  recursively if it does 
not exist.}

\item{aggregate_factor}{Integer. Aggregation factor expressed as number of
cells in each direction (horizontally and vertically). Or a vector of two
integers (horizontal and vertical aggregation factors) or three integers
(when also aggregating over layers).}

\item{fun}{Function. Function used to aggregate values. Default is
\code{sum}.}

\item{return_rast}{Logical. Should the resulting raster be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting raster is saved as a
  geotiff to that path. If \code{return_rast} is \code{TRUE} or
  \code{outfile} is not specified the resulting raster is returned,
  otherwise \code{NULL} is returned invisibly.
}
\description{
Aggregate raster cells (and optionally layers) to coarser resolution.
}
