% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphahull.R
\name{alphahull}
\alias{alphahull}
\title{Convert an alpha hull into an sf object}
\usage{
alphahull(xy, alpha)
}
\arguments{
\item{xy}{A two-column matrix of x, y coordinates.}

\item{alpha}{Alpha parameter for calculating the hull.}
}
\value{
A \code{\link{SpatVector}} object.
}
\description{
Convert an alphahull ahull object into an sf object.
}
\examples{
h <- alphahull(matrix(runif(100), ncol=2), 0.5)
plot(h, col=2)
}
\references{
Adapted from ConR:::.alpha.hull.poly and ConR:::.alpha.hull.poly.
}
\keyword{internal}
