% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals.R
\name{arrivals}
\alias{arrivals}
\title{Estimates pest arrivals}
\usage{
arrivals(
  weight,
  leakage,
  viability,
  outfile,
  return_rast = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{weight}{A \code{\link{SpatRaster}} or \code{\link{RasterLayer}} object, or a file path to a
raster dataset in a format supported by OGR, as produced by
\code{\link[=container_weights]{container_weights()}}.}

\item{leakage}{Numeric vector of 2 values, giving the lower and upper bounds
of a 95\% CI for leakage rate (the number of pest leakage events in a random
year).}

\item{viability}{Numeric vector of 2 values, giving the lower and upper
bounds of a 95\% CI for viability rate (the rate of survival, to
establishment, for leakage events).}

\item{outfile}{Character. Output raster file path. If \code{probability} has
length > 1, the file type must support multiband rasters (e.g. GeoTiff). If
not provided, the resulting \code{\link{SpatRaster}} object will be returned to R.}

\item{return_rast}{Logical. Should the \code{\link{SpatRaster}} object be returned to
R? Ignored if \code{outfile} is not provided.}

\item{overwrite}{Logical. Should \code{outfile} be overwritten if it exists?
Default is \code{FALSE}.}
}
\value{
If \code{outfile} is specified, the resulting \code{\link{SpatRaster}} object is
saved as a GeoTiff at that path. If \code{return_rast} is \code{TRUE} or \code{outfile} is
not specified, the resulting \code{\link{SpatRaster}} object is returned, otherwise
\code{NULL} is returned invisibly.
}
\description{
Estimates pest arrivals by containers.
}
\concept{functions estimating arrivals}
