% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals_by_containers.R
\name{arrivals_by_containers}
\alias{arrivals_by_containers}
\title{Estimates pest arrivals by containers}
\usage{
arrivals_by_containers(
  container_weights,
  port_data,
  template_raster,
  probability,
  outfile,
  summarise_uncertainty = FALSE,
  return_rast = FALSE
)
}
\arguments{
\item{container_weights}{An \code{\link{sf}} object or file path to a
shapefile supported by OGR, as produced by
\code{\link{container_weights}}.}

\item{port_data}{Character. Path to csv file containing Port Names, Port
Codes, Longitude, Latitude and Container volumes. Column names must be (in
this order): Name, PortCode, Longitude, Latitude, Count.}

\item{template_raster}{\code{RasterLayer} or file path to a raster file.
This is used to define the extent and resolution of output. Must be in
CRS EPSG:3577.}

\item{probability}{Numeric vector of one or more probabilities that a unit
carries a pest.}

\item{outfile}{Character. Output raster file path. If \code{probability} has
length > 1, the file type must support multiband rasters (e.g. GeoTiff). If
not provided, raster object will be returned to R.}

\item{return_rast}{Logical. Should the raster object be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting raster (multiband if
  \code{probability} has length > 1) is saved as a tiff at that path. If
  \code{return_rast} is \code{TRUE} or \code{outfile} is not specified the
  resulting raster object is returned, otherwise \code{NULL} is returned 
  invisibly.
}
\description{
Estimates pest arrivals by containers.
}
\seealso{
Other functions estimating arrivals: 
\code{\link{arrivals_by_fertiliser}()},
\code{\link{arrivals_by_goods}()},
\code{\link{arrivals_by_machinery}()},
\code{\link{arrivals_by_mail}()},
\code{\link{arrivals_by_nurserystock}()},
\code{\link{arrivals_by_residents}()},
\code{\link{arrivals_by_torres}()},
\code{\link{arrivals_by_tourists}()},
\code{\link{arrivals_by_vessels}()}
}
\concept{functions estimating arrivals}
