% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals_by_food.R
\name{arrivals_by_food}
\alias{arrivals_by_food}
\title{Estimates pest arrivals through imported fresh food}
\usage{
arrivals_by_food(
  pop_density,
  leakage_rate,
  establishment_rate,
  outfile,
  return_rast = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{pop_density}{A \code{RasterLayer} or file path to a raster file
containing population density.}

\item{leakage_rate}{Numeric vector of 2 values, giving the lower and upper
bounds of a 95% CI for leakage rate (the number of pest leakage events in
a random year).}

\item{establishment_rate}{Numeric vector of 2 values, giving the lower and
upper bounds of a 95% CI for establishment rate (the rate of survival, to
establishment, for leakage events).}

\item{outfile}{Character. Output raster file path. If \code{probability} has
length > 1, the file type must support multiband rasters (e.g. GeoTiff). If
not provided, raster object will be returned to R.}

\item{return_rast}{Logical. Should the raster object be returned to R?
Ignored if \code{outfile} is not provided.}

\item{overwrite}{Logical. If \code{TRUE} and \code{outfile} is not missing,
it will be overwritten if the file specified by \code{outfile} already
exists.}
}
\value{
If \code{outfile} is specified, the resulting raster (multiband if
  \code{probability} has length > 1) is saved as a tiff at that path. If
  \code{return_rast} is \code{TRUE} or \code{outfile} is not specified the
  resulting raster object is returned, otherwise \code{NULL} is returned
  invisibly.
}
\description{
Estimates pest arrivals through imported fresh food as a function of
population density.
}
\seealso{
Other functions estimating arrivals: 
\code{\link{arrivals_by_containers}()},
\code{\link{arrivals_by_fertiliser}()},
\code{\link{arrivals_by_goods}()},
\code{\link{arrivals_by_machinery}()},
\code{\link{arrivals_by_mail}()},
\code{\link{arrivals_by_nurserystock}()},
\code{\link{arrivals_by_residents}()},
\code{\link{arrivals_by_torres}()},
\code{\link{arrivals_by_tourists}()},
\code{\link{arrivals_by_vessels}()}
}
\concept{functions estimating arrivals}
