% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals_by_mail.R
\name{arrivals_by_mail}
\alias{arrivals_by_mail}
\title{Estimates pest arrivals through mail}
\usage{
arrivals_by_mail(pop_density, total_mail, probability, outfile,
  return_rast = FALSE)
}
\arguments{
\item{pop_density}{A \code{RasterLayer} or file path to a raster file 
containing population density.}

\item{total_mail}{Integer. Amount of mail.}

\item{probability}{Numeric. The probability a mail item contains pest.}

\item{outfile}{Character. Output raster file path. If not provided, raster 
object will be returned to R.}

\item{return_rast}{Logical. Should the raster object be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting raster is saved as a
  to that path. If \code{return_rast} is \code{TRUE} or \code{outfile} is 
  not specified the resulting raster is returned, otherwise \code{NULL} is 
  returned invisibly.
}
\description{
Estimates arrival rate attributable to mail as a function of population
density.
}
\seealso{
Other functions estimating arrivals: \code{\link{arrivals_by_containers}},
  \code{\link{arrivals_by_fertiliser}},
  \code{\link{arrivals_by_food}},
  \code{\link{arrivals_by_machinery}},
  \code{\link{arrivals_by_nurserystock}},
  \code{\link{arrivals_by_residents}},
  \code{\link{arrivals_by_torres}},
  \code{\link{arrivals_by_tourists}},
  \code{\link{arrivals_by_vessels}}
}
\concept{functions estimating arrivals}
