% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals_by_nurserystock.R
\name{arrivals_by_nurserystock}
\alias{arrivals_by_nurserystock}
\title{Estimates pest arrivals through imported nursery stock}
\usage{
arrivals_by_nurserystock(
  pop_density,
  total_imports,
  probability,
  outfile,
  return_rast = FALSE
)
}
\arguments{
\item{pop_density}{A \code{RasterLayer} or file path to a raster file 
containing population density.}

\item{total_imports}{Integer. Amount of imported nursery stock entering
country.}

\item{probability}{Numeric. The probability a imported nursery stock item 
contains pest.}

\item{outfile}{Character. Output raster file path. If not provided, raster 
object will be returned to R.}

\item{return_rast}{Logical. Should the raster object be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting raster is saved as a
  to that path. If \code{return_rast} is \code{TRUE} or \code{outfile} is 
  not specified the resulting raster is returned, otherwise \code{NULL} is 
  returned invisibly.
}
\description{
Estimates pest arrivals through imported nursery stock as a function of
population density.
}
\seealso{
Other functions estimating arrivals: 
\code{\link{arrivals_by_containers}()},
\code{\link{arrivals_by_fertiliser}()},
\code{\link{arrivals_by_goods}()},
\code{\link{arrivals_by_machinery}()},
\code{\link{arrivals_by_mail}()},
\code{\link{arrivals_by_residents}()},
\code{\link{arrivals_by_torres}()},
\code{\link{arrivals_by_tourists}()},
\code{\link{arrivals_by_vessels}()}
}
\concept{functions estimating arrivals}
