% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals_by_vessels.R
\name{arrivals_by_vessels}
\alias{arrivals_by_vessels}
\title{Estimates pest arrivals by vessels}
\usage{
arrivals_by_vessels(
  port_weight,
  n_vessels,
  probability,
  outfile,
  summarise_uncertainty = FALSE,
  return_rast = FALSE
)
}
\arguments{
\item{port_weight}{A \code{RasterLayer} or file path to a raster file 
containing weights to be used to distribute arrivals.}

\item{n_vessels}{Integer. The total number of vessels entering country.}

\item{probability}{Numeric vector of one or more probabilities that a unit
carries a pest.}

\item{outfile}{Character. Output raster file path. If \code{probability} has
length > 1, the file type must support multiband rasters (e.g. GeoTiff). If
not provided, raster object will be returned to R.}

\item{return_rast}{Logical. Should the raster object be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting raster (multiband if
  \code{probability} has length > 1) is saved as a tiff at that path. If
  \code{return_rast} is \code{TRUE} or \code{outfile} is not specified the
  resulting raster object is returned, otherwise \code{NULL} is returned 
  invisibly.
}
\description{
Estimates pest arrivals by vessels.
}
\seealso{
Other functions estimating arrivals: 
\code{\link{arrivals_by_containers}()},
\code{\link{arrivals_by_fertiliser}()},
\code{\link{arrivals_by_goods}()},
\code{\link{arrivals_by_machinery}()},
\code{\link{arrivals_by_mail}()},
\code{\link{arrivals_by_nurserystock}()},
\code{\link{arrivals_by_residents}()},
\code{\link{arrivals_by_torres}()},
\code{\link{arrivals_by_tourists}()}
}
\concept{functions estimating arrivals}
