% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize_and_aggregate.R
\name{binarize_and_aggregate}
\alias{binarize_and_aggregate}
\title{Binarize a categorical raster and optionally aggregate}
\usage{
binarize_and_aggregate(
  infile,
  rle,
  outfile,
  extent,
  res,
  categories,
  overwrite = FALSE,
  return_rast = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{infile}{Character. File path to a categorical raster. If \code{rle} is
provided, \code{infile} must not be provided.}

\item{rle}{A \code{raster_rle} object generated by \code{\link[=rle_compress]{rle_compress()}} or a file path
to such an object saved as .rds.}

\item{outfile}{Character. The target file path for the binarized raster.}

\item{extent}{Output extent as a \code{\link{SpatExtent}} object or an object from
which a \code{\link{SpatExtent}} object can be extracted. If not provided, extent
will be taken from input. If \code{extent} differs from that of \code{infile}, new
cells will be assigned value 1 if any original cells belonging to
\code{categories} have their centroids within the new cell.}

\item{res}{output resolution as a numeric vector of 1 or 2 elements, or a
\code{\link{SpatRaster}} or \verb{Raster*} object from which resolution can be extracted.
If not provided, resolution will be taken from input. If \code{res} differs from
that of \code{infile}, new cells will be assigned value 1 if any original cells
belonging to \code{categories} have their centroids within the new cell.}

\item{categories}{Integer or numeric vector of class values to be laballed as
1 in the target raster.}

\item{overwrite}{Logical. Should \code{outfile} be overwritten if it exists?}

\item{return_rast}{Logical. Should the target \code{\link{SpatRaster}} be returned to R
(\code{TRUE}), or not returned (\code{FALSE})?}

\item{quiet}{Logical. Should progress messages be suppressed?}
}
\value{
A binarized raster layer is written to \code{outfile}, and if
\code{return_rast} is \code{TRUE}, the raster is additionally returned to R
as a \code{\link{SpatRaster}} layer. If \code{outfile} is provided and \code{return_rast} is
\code{FALSE}, then \code{outfile} is returned invisibly.
}
\description{
Binarize a categorical raster and optionally aggregate and/or resample.
}
