% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabi_flagger.R
\name{cabi_flagger}
\alias{cabi_flagger}
\title{Flags records that occur in countries that are not reported in CABI}
\usage{
cabi_flagger(occurrence_records, cabi_ref, manual_check = FALSE,
  return_df = FALSE)
}
\arguments{
\item{occurrence_records}{A \code{data.frame}, \code{sf} object or
\code{SpatialPoints*} object, or a path to a .csv file containing the 
locations of species of interest. (If a \code{data.frame} is provided,
it must contain the named columns "Latitude" and "Longitude").}

\item{cabi_ref}{Character. The path to a .csv file downloaded from a CABI
datasheet containing the country-scale distributional data for the pest.}

\item{manual_check}{Logical. Allows interactive selection of which 
unverified points to retain. Interactive map will only be produced if 
unverified records are present and if this argument is set to \code{TRUE}.}

\item{return_df}{Logical. Return the object as a \code{data.frame} or as an 
\code{sf} object (i.e. \code{TRUE}, the default).}
}
\value{
A \code{data.frame} or an \code{sf} points object flagging records 
  found in countries in which CABI does not report the species as being 
  established.
}
\description{
Flags records that occur in countries that are not reported in CABI.
}
