% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EE.R
\name{calc_EE}
\alias{calc_EE}
\title{Calculate establishment probability}
\usage{
calc_EE(n_events, p_establish, nsims = 1e+05)
}
\arguments{
\item{n_events}{Numeric vector containing lower and upper bounds (95\% CI) for
the number of leakage events.}

\item{p_establish}{Numeric vector containing the lower and upper bounds (95\%
CI) for the probability a leakage event could result in an establishment.}

\item{nsims}{Integer. Number of samples to be taken from event and
establishment distributions. Default = 100000.}
}
\value{
A \code{data.frame} containing the possible number of incursions that
may occur, and their corresponding probabilities.
}
\description{
Calculate the probability of one or more establishment events occurring as as
a function of estimated leakage numbers and the probability a leakage event
could result in establishment.
}
\author{
James Camac (\email{james.camac@gmail.com})
}
