% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_proportion.R
\name{calc_proportion}
\alias{calc_proportion}
\title{Calculate raster cell values as proportions}
\usage{
calc_proportion(rast)
}
\arguments{
\item{rast}{Raster object.}
}
\value{
A \code{RasterLayer}.
}
\description{
Calculates raster cell values as a proportion of the sum of all cells' values.
}
