% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_extent.R
\name{check_extent}
\alias{check_extent}
\title{Check and adjust the extent of a raster}
\usage{
check_extent(extent, res, snap = c("out", "in"))
}
\arguments{
\item{extent}{The extent to check.}

\item{res}{The resolution to check against.}

\item{snap}{The direction to snap the extent. Can be "out" or "in".}
}
\value{
The adjusted extent.
}
\description{
This function checks whether longitudinal and latitudinal ranges are
multiples of \code{res}, and if not snaps the extent in or out according to
\code{snap}.
}
\examples{
check_extent(extent = c(0, 1, 0, 1), res = 0.5, snap = "out")
check_extent(extent = c(0, 1, 0, 1), res = 0.3, snap = "out")
check_extent(extent = c(0, 1, 0, 1), res = 0.3, snap = "in")
}
