% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_weight.R
\name{container_weight}
\alias{container_weight}
\title{Creates a weight by postcode for each major port}
\usage{
container_weight(
  path,
  sheet_nums,
  range = "A7:M2217",
  postcodes,
  port_data,
  na = c("", "-", "np"),
  template_raster,
  outfile,
  return_rast = FALSE
)
}
\arguments{
\item{path}{Character. File path to Microsoft Excel .xls file containing
containers by postcode for each port.}

\item{sheet_nums}{Integer. Vector of integers signifying the sheet numbers to
read in.}

\item{range}{A cell range to read from, as described in cell-specification.
Includes typical \emph{Excel} ranges such as \code{"B3:D87"}, possibly including the
sheet name like \code{"Budget!B2:G14"}, and more. Interpreted strictly, even if
the range forces the inclusion of leading or trailing empty rows or
columns.}

\item{postcodes}{Character. File path to postcode vector dataset (e.g. ESRI
Shapefile).}

\item{port_data}{Character. Path to csv file containing Port Names, Port
Codes, Longitude, Latitude and Container volumes. Column names must be (in
this order): Name, PortCode, Longitude, Latitude, Count.}

\item{na}{Character vector of strings to interpret as missing values. By
default, \code{readxl} treats blank cells as missing data.}

\item{template_raster}{\code{\link{SpatRaster}} or \code{RasterLayer} object, or a file
path to a raster file. This is used to define the extent and resolution of
output. Must be specified in CRS EPSG:3577.}

\item{outfile}{Character. Name of shapefile where output will be saved. If
not provided, \code{\link{SpatVector}} object will be returned to R.}

\item{return_rast}{Logical. Should the \code{\link{SpatRaster}} object be returned to
R? Ignored if \code{outfile} is not provided.}
}
\value{
A \code{\link{SpatRaster}} object if \code{return_rast} is \code{TRUE} or if \code{outfile} is
missing; otherwise, \code{outfile} is returned invisibly.
}
\description{
Creates a weight by postcode for each major port.
}
\details{
For the purposes of this analysis missing container counts (i.e.
NAs) are treated as zeroes.
}
