% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disperse_arrivals.R
\name{disperse_arrivals}
\alias{disperse_arrivals}
\title{Disperse pest arrivals}
\usage{
disperse_arrivals(rast, n_vectors, probability)
}
\arguments{
\item{rast}{Raster object containing dispersal weights.}

\item{n_vectors}{Numeric. The number of transmission vectors (e.g.
passengers).}

\item{probability}{Numeric. The probability a vector contains pest.}
}
\value{
A raster object.
}
\description{
Disperses pest arrivals based on input raster.
}
