% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edmaps-package.R
\docType{package}
\name{edmaps-package}
\alias{edmaps}
\alias{edmaps-package}
\title{edmaps: Estimate Likelihood of Pest Establishment}
\description{
This package combines spatial environmental data (e.g. distance from airports/marine ports, distribution of pest habitat/hosts) with biosecurity leakage rates (e.g. expert-elicited) to generate maps of the likelihood of pest establishment across the landscape. Final outputs include raster datasets (GTiff) indicating establishment likelihood across the area of interest, as well as interactive (html) and static (pdf) versions of these maps.
}
\author{
\strong{Maintainer}: James Camac \email{james.camac@gmail.com}

Authors:
\itemize{
  \item John Baumgartner \email{johnbaums@gmail.com}
}

}
\keyword{internal}
