% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_function.R
\name{exp_function}
\alias{exp_function}
\title{Exponentiate raster values}
\usage{
exp_function(rast, beta)
}
\arguments{
\item{rast}{\code{\link{SpatRaster}} object.}

\item{beta}{Numeric. The beta coefficient of a standard exponential function.}
}
\value{
A \code{\link{SpatRaster}} object.
}
\description{
Apply exponential function to raster values.
}
