% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_range}
\alias{expand_range}
\title{Expand sequences}
\usage{
expand_range(x)
}
\arguments{
\item{x}{A character vector of shorthand sequences described by colons (e.g.
\code{c('1:3', '5:10')}) or hyphens (e.g. \code{c('1-3', '5-10'}).}
}
\value{
Expanded integer sequences.
}
\description{
Expand sequence specified with colon or hyphen.
notation.
}
\keyword{internal}
