% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertiliser_by_nrm.R
\name{fertiliser_by_nrm}
\alias{fertiliser_by_nrm}
\title{Create fertiliser by nrm sf object}
\usage{
fertiliser_by_nrm(abs_data, nrm_shapefile, outfile, return_vect = FALSE)
}
\arguments{
\item{abs_data}{Character. File path to ABS .csv file.}

\item{nrm_shapefile}{Character. File path to NRM shapefile.}

\item{outfile}{Character. Name of shapefile (or other vector data format
supported by OGR) where output will be saved. If not provided, \code{sf} object
will be returned to R.}

\item{return_vect}{Logical. Should the \code{\link{SpatVector}} object be returned to
R? Ignored if \code{outfile} is not provided.}
}
\value{
If \code{return_vect} is \code{TRUE} or \code{outfile} is missing, a \code{\link{SpatVector}}
object will be returned. Otherwise, the vector dataset is written to
\code{outfile} and NULL is returned invisibly.
}
\description{
Create fertiliser by nrm sf object.
}
