% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertiliser_weight.R
\name{fertiliser_weight}
\alias{fertiliser_weight}
\title{Create a fertiliser weight raster}
\usage{
fertiliser_weight(fert_nrm, fert_landuses, outfile, return_rast = FALSE)
}
\arguments{
\item{fert_nrm}{A file path to vector data or an \code{sf} object.}

\item{fert_landuses}{A file path to raster file or  a \code{RasterLayer}
object.}

\item{outfile}{Character. Output raster file path. If not provided,
\code{RasterLayer} will be returned to R.}

\item{return_rast}{Logical. Should the \code{RasterLayer} be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting \code{RasterLayer} is
saved as to that path. If \code{return_rast} is \code{TRUE} or
\code{outfile} is not specified, the resulting \code{RasterLayer} is
returned, otherwise \code{NULL} is returned invisibly.
}
\description{
Create a fertiliser weight raster as a function of estimated nrm fertiliser
tonnes and landuses.
}
