% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excel_globals.R
\name{get_excel_globals}
\alias{get_excel_globals}
\title{Get global parameters from Excel file}
\usage{
get_excel_globals(file)
}
\arguments{
\item{file}{The path to the Excel file.}
}
\value{
A list of global parameters.
}
\description{
This function reads the "Global parameters" sheet from the specified Excel file and returns a list of global parameters.
}
\examples{
get_excel_globals(
  system.file("extdata", "parameters.xlsx", package = "edmaps")
)
}
