% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excel_pathways.R
\name{get_excel_pathways}
\alias{get_excel_pathways}
\title{Get pathway data from an Excel file}
\usage{
get_excel_pathways(file)
}
\arguments{
\item{file}{The path to the Excel file.}
}
\value{
A list of pathways.
}
\description{
This function reads the "Pathways" sheet from the specified Excel file and
returns a list of pathways.
}
\examples{
get_excel_pathways(
  system.file("extdata", "pathways.xlsx", package = "edmaps")
)
}
