% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excel_species.R
\name{get_excel_species}
\alias{get_excel_species}
\title{Get species data from an Excel file}
\usage{
get_excel_species(file)
}
\arguments{
\item{file}{The path to the Excel file.}
}
\value{
A data frame containing the cleaned species data.
}
\description{
This function reads in an Excel file and extracts the species data from the
"Species" sheet. It then renames the columns to match the expected input
format, and performs some data cleaning. Finally, it returns the cleaned
species data.
}
\examples{
get_excel_species(
  system.file("extdata", "parameters.xlsx", package = "edmaps")
)
}
