% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_map.R
\name{interactive_map}
\alias{interactive_map}
\title{Produce an interactive html map}
\usage{
interactive_map(ras, layer_name = NULL, palette = "inferno",
  transparency = 0.8, legend = TRUE, set_value_range = NULL,
  discrete = FALSE, scale_type = "none", outfile = NULL,
  surveillance_locs = NULL, pt_col = "red", cleanup = FALSE)
}
\arguments{
\item{ras}{A \code{RasterLayer} or file path to a raster file.}

\item{layer_name}{Character. An optional name to assign to \code{ras}.}

\item{palette}{Either a vector of 2 or more colours (e.g. as hex codes or
colour names) or the name of a palette function supported by \code{tmap}
(see \code{\link[tmaptools]{palette_explorer}} and
\code{\link[tmap]{tm_raster}}).}

\item{transparency}{Numeric. Value between 0 and 1 defining the opacity of 
the plotted raster data (1 = fully opaque; 0 = fully transparent).}

\item{legend}{Logical. Should a legend be plotted?}

\item{set_value_range}{A numeric vector giving upper and lower limits for
raster values. Values outside this range (including the limits) will be 
set to NA.}

\item{discrete}{Logical. Are the values of \code{ras} discrete 
(categorical)?}

\item{scale_type}{Character. Can be one of: \code{"none"} (raw data, no
rescaling), \code{"log10"}, \code{"max normalize"} (proportional to 
maximum value), \code{"minmax normalize"} (rescale values to be between 0 
and 1 based on min and max) or \code{"logit"}. Note that if 
\code{"log10"} or \code{"logit"} is used, 0 or 1 values must be masked (
using \code{set_value_range}) or rescaled before passing to this 
function. \code{scale_type} is ignored if \code{discrete} is \code{TRUE}.}

\item{outfile}{Character. If \code{NULL}, map will be returned to R and not
saved. Otherwise, map will be exported as a html file. Full path address
must be used. If pandoc is available, a standalone html file is created
(see details).}

\item{surveillance_locs}{A spatial object or a path to a .csv file 
containing columns named "Latitude" and "Longitude".}

\item{pt_col}{Character. The plotting colour for surveillance points.}

\item{cleanup}{Logical. If a standalone html file is created, should 
accessory files be removed after the standalone file is generated? This 
will be a folder created within \code{tempdir()}.}
}
\value{
A html map.
}
\description{
Produce an interactive html map.
}
\details{
To create a standalone html file, the 
  \href{https://pandoc.org/installing.html}{pandoc} software must be 
  installed and available to R. If pandoc is unavailable, the html file 
  will be accompanied by a folder of accessory files.
}
