% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edmaps.R
\name{make_edmaps}
\alias{make_edmaps}
\title{Make Plan}
\usage{
make_edmaps(file)
}
\arguments{
\item{file}{The path to the Excel file.}
}
\value{
A list of targets for the pipeline.
}
\description{
This function generates a plan for the target pipeline.
}
\examples{
make_edmaps(
  system.file("extdata", "parameters.xlsx", package = "edmaps")
)
}
