% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_mask.R
\name{na_mask}
\alias{na_mask}
\title{Create a mask raster}
\usage{
na_mask(infile, outfile, res, extent)
}
\arguments{
\item{infile}{Character. File path to input raster file.}

\item{outfile}{Character. Output raster file path. Directory will be created
(recursively) if it doesn't exist.}

\item{res}{Numeric. A vector of one or two numbers giving the desired output
resolution. If missing, the input resolution will be used. If a single
number is provided, it will be used for both horizontal and vertical
resolution.}

\item{extent}{One of of: a numeric vector giving the desired extent of the
output raster as \code{c(xmin, xmax, ymin, ymax)}; a \code{\link{SpatExtent}} object; or
an object from which such an object can be extracted.}
}
\value{
A raster is written to \code{outfile}, with NA cells transferred from
\code{infile} and all other cells set to 1. \code{outfile} is returned invisibly.
}
\description{
Create a mask raster, retaining NA and setting all non-NA cells to 1. Input
raster will be resampled if necessary, using "max" resampling (i.e. if any
contributing cells are not NA, the new cell will be set to 1).
}
