% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncells_4_threshold.R
\name{ncells_4_threshold}
\alias{ncells_4_threshold}
\title{Extract number of cells required to meet threshold proportion of risk}
\usage{
ncells_4_threshold(risk_rasters, names, proportion_captured = c(0.6, 0.8,
  0.9, 0.95))
}
\arguments{
\item{risk_rasters}{Character. File path(s) to rasters to be loaded.}

\item{names}{Character. Names corresponding to \code{infiles}.}

\item{proportion_captured}{Numeric vector. Proportion(s) of risk to be 
captured.}
}
\value{
A \code{data.frame} containing the number of cells to be trapped to
  capture given proportion(s) of total risk across multiple input files.
}
\description{
Extract number of cells required to meet threshold proportion of risk.
}
