% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_establishment_captured.R
\name{plot_establishment_captured}
\alias{plot_establishment_captured}
\title{Plots the cumulative proportion of establishment likelihood captured}
\usage{
plot_establishment_captured(
  data,
  xlab = "Number of cells",
  ylab = "Proportion",
  legend_title = NULL,
  legend = "right",
  prop_line,
  y_limit = NULL,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm"),
  outfile
)
}
\arguments{
\item{data}{A \code{data.frame} as derived from \code{\link[=captured_by_ncells]{captured_by_ncells()}}.}

\item{xlab}{Character. x axis label.}

\item{ylab}{Character. y axis label.}

\item{legend_title}{Character. Optional legend title.}

\item{legend}{Legend position. Either a position name (one of
\code{"none"}, \code{"top"}, \code{"bottom"}, \code{"left"}, or
\code{"right"}, or a vector of two normalised coordinates ranging from 0
to 1, e.g. \code{c(1, 0)} (for bottom-right).}

\item{prop_line}{Numeric. Proportion in which to add a vertical line.
Default is NULL (no line).}

\item{y_limit}{Numeric vector giving the minimum and maximum y-axis limits.
If omitted this will be determined based on the data.}

\item{width}{Width of plot. If not defined will use size of current graphic
device.}

\item{height}{Height of plot. If not defined will use size of current
graphic device.}

\item{units}{Character. Units corresponding to \code{height} and
\code{width}. Can be \code{"in"}, \code{"cm"}, or \code{"mm"}. Default is
inches (\code{"in"}).}

\item{outfile}{Character. Output image file path. Containing directory
will be created recursively if it does not already exist.}
}
\value{
An image is written to \code{outfile} if provided, and otherwise a
\code{ggplot} object is returned.
}
\description{
Plots the cumulative proportion of establishment likelihood captured as a
greater number of top ranking risk cells is considered.
}
