% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_raster}
\alias{plot_raster}
\title{Plot a raster}
\usage{
plot_raster(
  rast,
  legend_title,
  occurrence_data = NULL,
  pt_col = "red",
  height,
  compass = FALSE,
  outfile
)
}
\arguments{
\item{rast}{A \code{RasterLayer}, single-layer \code{\link{SpatRaster}} or file path to
raster file.}

\item{legend_title}{Character. If missing, the name of the raster layer will
be used.}

\item{occurrence_data}{A \code{data.frame}, \code{sf} object, \code{SpatialPointsDataFrame}
object, \code{\link{SpatVector}} object, or path to a .csv file containing columns
named "Latitude" and "Longitude". If \code{NULL}, no points will be plotted.}

\item{pt_col}{Character. Colour of points (if plotted).}

\item{height}{Height of plot in inches (will be rendered at 300 dpi).
Required if \code{outfile} is provided.}

\item{compass}{Logical. Should a North arrow be shown?}

\item{outfile}{Character. Path to save output.}
}
\value{
A \code{tmap} object. If \code{outfile} is provided, a map will also be written
to that file.
}
\description{
Plot a raster.
}
