% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointInPoly}
\alias{pointInPoly}
\title{Check if a point is contained within a polygon}
\usage{
pointInPoly(point, vertices)
}
\arguments{
\item{point}{A \code{rowvec} with x,y coordinate structure.}

\item{vertices}{A \code{matrix} containing the boundary points of the polygon.}
}
\value{
A \code{bool} indicating whether the point is in the polygon (TRUE) or not (FALSE)
}
\description{
Check if a point is contained within a polygon
}
