% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointsInPoly}
\alias{pointsInPoly}
\title{Check if multiple points are contained within a polygon}
\usage{
pointsInPoly(points, vertices)
}
\arguments{
\item{points}{A \code{matrix} with x,y coordinate structure.}

\item{vertices}{A \code{field} of type {matrix} that contains the polygon coordinates to test against.}
}
\value{
A logical \code{vector} indicating whether each point is within the polygon.
}
\description{
Check if multiple points are contained within a polygon
}
