% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_establishment.R
\name{pr_establishment}
\alias{pr_establishment}
\title{Calculate probability of establishment}
\usage{
pr_establishment(rast, suitability, outfile, return_rast)
}
\arguments{
\item{rast}{\code{RasterLayer}, single-layer \code{\link{SpatRaster}}, or path to a raster
file containing pathway probability of pest arrival.}

\item{suitability}{\code{RasterLayer}, single-layer \code{\link{SpatRaster}}, or path to a
raster file containing a suitability raster. Cell values must be between 0
(not suitable) and 1 (ideal suitability).}

\item{outfile}{Character. Output raster file path. If not provided, the
\code{\link{SpatRaster}} will be returned to R.}

\item{return_rast}{Logical. Should the \code{\link{SpatRaster}} be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting \code{\link{SpatRaster}} is saved to
\code{outfile}. If \code{return_rast} is \code{TRUE} or \code{outfile} is not specified, the
resulting \code{\link{SpatRaster}} is returned, otherwise \code{NULL} is returned
invisibly.
}
\description{
Calculate probability of establishment for each raster cell as a function of
probabilities of arrival and climate and/or biotic suitability.
}
