% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_weight.R
\name{proximity_weight}
\alias{proximity_weight}
\title{Create a seaport proximity weighting raster}
\usage{
proximity_weight(
  pts,
  rast,
  pt_weight,
  cell_weight,
  beta,
  method = c("nearest", "all"),
  outfile,
  return_rast = FALSE
)
}
\arguments{
\item{pts}{\code{\link{SpatVector}} with point geometry, or a filepath to a spatial
dataset with point geometry. CRS must be defined.}

\item{rast}{\verb{Raster*} or \code{\link{SpatRaster}} object, or a file path to template
raster. The template raster will also be used to mask the output raster so
that NA cell values will be propagated. CRS must be defined. Distance will
be calculated for all non-NA cells.}

\item{pt_weight}{An optional vector of weights that correspond to elements of
\code{pts}. These weights are applied after the negative exponential distance
weighting, by multiplication. E.g., if \code{pts} is a set of seaports,
\code{pt_weight} might be a vector of annual vessel counts for each port. When
\code{method='nearest'}, length of \code{pt_weight} must be 1. When \code{method='all'},
length of \code{pt_weight} can be 1 (recycled to length of \code{pts}), or the number
of features in \code{pts}.}

\item{cell_weight}{An optional \code{RasterLayer} or \code{\link{SpatRaster}} object, or
path to a raster file, containing cell weights to apply to the weighted
distances. E.g., if \code{pts} is a set of airports, \code{cell_weight} could be a
raster of human population density. Must have same geometry as \code{rast}.}

\item{beta}{Numeric. Parameter passed to the exponential function. Distance
to nearest airport is multiplied by this value and exponentiated to give
the relative density of tourists at a location. To generate a distribution
that ensures proportion \emph{p} of tourists within distance \emph{d} of nearest
airport, specify \code{beta=log(p)/d} (e.g. to have 50\% of tourists
within 200 km of an airport, use \code{log(0.5)/200000}) (assuming that either
the map is in long/lat or map units are metres).}

\item{method}{Either \code{'nearest'} (default; only consider distance to the
nearest point in \code{pts}) or \code{'all'} (sum weighted distances to all points in
\code{pts}).}

\item{outfile}{Character. Output raster file path. If not provided, the
resulting \code{\link{SpatRaster}} will be returned to R.}

\item{return_rast}{Logical. Should the \code{\link{SpatRaster}} be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting \code{\link{SpatRaster}} is saved as
to that path. If \code{return_rast} is \code{TRUE} or \code{outfile} is not specified, the
resulting \code{\link{SpatRaster}} is returned, otherwise \code{outfile} is returned
invisibly. Note that cell weights are divided by their global sum.
}
\description{
Create a weight raster describing distance from one or more points, with
negative exponential distance decay.
}
