% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_vector.R
\name{rasterize_vector}
\alias{rasterize_vector}
\title{Rasterize spatial vector dataset}
\usage{
rasterize_vector(
  vector_data,
  outfile,
  template_raster,
  extent,
  res,
  crs,
  field,
  burn_value,
  datatype = "Float32",
  overwrite = FALSE,
  return_rast = FALSE
)
}
\arguments{
\item{vector_data}{Character. File path to vector spatial data.}

\item{outfile}{Raster output file path. Parent directory will be created
recursively if required.}

\item{template_raster}{Optional. \verb{Raster*}, \code{\link{SpatRaster}}, or a file path to
template raster. If this is provided, \code{extent}, \code{res}, and \code{crs} will be
taken from this raster unless they are also passed to this function. If \code{x}
is not provided, then \code{extent} and \code{res} must be provided.}

\item{extent}{Either a character path to a raster file, an
\link[terra:SpatExtent-class]{SpatExtent} object (or an object from which such an
extent can be extracted), or a numeric vector with four elements giving
xmin, xmax, ymin, ymax.}

\item{res}{Numeric or integer vector giving the horizontal and vertical
spatial resolution of the target raster, in units of \code{crs}. If a single
value is given, it will be used for both horizontal and vertical
resolution.}

\item{crs}{Target coordinate reference system as a PROJ string (character) an
object of class CRS. If missing and \code{template_raster} is supplied, the crs
of \code{template_raster} will be used. If \code{template_raster} is also not
supplied, the CRS of \code{vector_data} will be used.}

\item{field}{Character. Name of attribute to be burned into raster.}

\item{burn_value}{Numeric. A constant value to burn into raster.}

\item{datatype}{Character. Output data type (see
[gdal_rasterize(https://gdal.org/programs/gdal_rasterize.html)
documentation).}

\item{overwrite}{Logical. Should outfile be overwritten if it exists?}

\item{return_rast}{Logical. Return object to R?}
}
\value{
A binarized raster is written to \code{outfile}, and returned to R as a
\code{\link{SpatRaster}} if \code{return_rast} is \code{TRUE}.
}
\description{
Converts vector object to a raster with specified extent and resolution.
}
