% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_flagger.R
\name{record_flagger}
\alias{record_flagger}
\title{Flags records that occur in countries that are not reported in either
infected_countries or CABI datasheet}
\usage{
record_flagger(
  occurrence_records,
  infected_countries,
  cabi_ref,
  return_df = FALSE
)
}
\arguments{
\item{occurrence_records}{A \code{data.frame}, \code{sf} object or \verb{SpatialPoints*}
object, or a path to a .csv file containing the locations of species of
interest. (If a \code{data.frame} is provided, it must contain the named columns
"Latitude" and "Longitude").}

\item{infected_countries}{Character vector of countries with known established
populations.}

\item{cabi_ref}{Character. The path to a .csv file downloaded from a CABI
datasheet containing the country-scale distributional data for the pest.
Note that if infected_countries the CABI data will not be used.}

\item{return_df}{Logical. Return the object as a \code{data.frame} or as an \code{sf}
object (i.e. \code{TRUE}, the default).}
}
\value{
A \code{data.frame} or an \code{sf} points object flagging records
found in countries in which do not occur in infected_countries OR
is not reported as being present in CABI.
}
\description{
Cross reference occurrence records against either a vector of country names
with established populations specified countries, or those presented in a
CABI distribution file. Occurrence records that occur in countries outside
one of these lists are then removed.
}
