% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_compress.R
\name{rle_compress}
\alias{rle_compress}
\title{Compress raster data using run length encoding}
\usage{
rle_compress(x, outfile, quiet = FALSE)
}
\arguments{
\item{x}{File path to the categorical raster to be compressed, or a
\code{Raster*} object.}

\item{outfile}{Character (optional). Path to target .rds file that will 
store RLE results. Directory will be created recursively if it doesn't 
exist.}

\item{quiet}{Logical. Should progress messages be suppressed?}
}
\value{
A list with five elements:
\describe{
  \item{starts}{Cell numbers corresponding to run starts}
  \item{lengths}{Run lengths}
  \item{values}{Run values}
  \item{extent}{Raster extent}
  \item{res}{Raster resolution}
}
This object is additionally saved in rds format to \code{outfile}, if
provided.
}
\description{
Compress categorical raster data using run length encoding.
}
