% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_map.R
\name{static_map}
\alias{static_map}
\title{Produce a static map of establishment likelihood}
\usage{
static_map(
  ras,
  xlim,
  ylim,
  subset_layers,
  layer_names,
  legend_title,
  set_value_range,
  scale_type = "none",
  basemap_mode = c("osm", "boundaries"),
  transparency = 0.7,
  colramp_entire_range = TRUE,
  surveillance_locs,
  shape = 21,
  pt_col = "red",
  aggregate_raster,
  nrow,
  height,
  outfile
)
}
\arguments{
\item{ras}{A \verb{Raster*}, \code{\link{SpatRaster}}, or file paths to one or more raster
files.}

\item{xlim}{Numeric vector. The longitudinal extent of the area to plot, in
WGS84 coordinates.}

\item{ylim}{Numeric vector. The latitudinal extent of the area to plot, in
WGS84 coordinates.}

\item{subset_layers}{Vector. A vector of layer names or numeric positions to
be included in plot. Only relevant if \code{ras} has multiple layers/bands. If
missing all layers are included.}

\item{layer_names}{Character vector. A vector of panel titles (only relevant
if ras is a stack). If not provided, names from ras will be used. If
provided, length must equal \code{terra::nlyr(ras)}, or match
\code{length(subset_layers)}.}

\item{legend_title}{Character. Legend title.}

\item{set_value_range}{A numeric vector containing an upper and lower bound
value (in units of raster). Values outside this range (including values
falling on the boundaries) will be masked.}

\item{scale_type}{Character. One of:
\itemize{
\item \code{"none"} (raw data, no rescaling);
\item \code{"log10"};
\item \code{"max normalize"} (proportional to maximum value);
\item \code{"minmax normalize"} (rescale values to be between zero and 1 based
on min and max); or
\item \code{"discrete"}.
Note that if \code{"log10"} is used 0 or 1 values. must be masked
(using \code{set_value_range}) or rescaled outside of this function.
}}

\item{basemap_mode}{Either \code{'boundaries'} or \code{'osm'} (i.e., OpenStreetMap),
defining whether OpenStreetMap imagery should be used for static map
basemaps, or simple administrative boundaries. Default is \code{'osm'}.}

\item{transparency}{Numeric. Transparency of raster, between 0-1.}

\item{colramp_entire_range}{Logical. Whether to set colour ramp limits based
on national risk range (\code{TRUE}) or by risk range present in region
specified by \code{xlim} and \code{ylim}.}

\item{surveillance_locs}{A spatial object or a path to a .csv file containing
columns named "Latitude" and "Longitude".}

\item{shape}{shape(s) to use for surveillance locations. Default is 21. If
you have multiple bands, and you'd like to use different symbols on each
panel, specify a vector of shapes. (e.g. c(NA, 21,21,21) will not include
shapes on the first panel, but will on the rest}

\item{pt_col}{Character. The plotting colour for surveillance points.}

\item{aggregate_raster}{\code{NULL} or a list containing the aggregation factor
(i.e. number of raster cells to aggregate) and the aggregation function
e.g. \code{list(10, sum)}.}

\item{nrow}{For multipanel plots, an optional  numeric value defining the
number of rows of panels.}

\item{height}{height of plot in inches (will be rendered at 300 dpi). If not
defined will use size of current graphic device. Width will be determined
automatically, based on the aspect ratio given by the plotting extent.}

\item{outfile}{Character. File path to save map.}
}
\value{
If \code{outfile} is provided, a map is saved to that file. Otherwise, a
\code{tmap} object is returned to R.
}
\description{
Produce a static map of establishment likelihood, with OpenStreetMap base
layer.
}
\details{
This function relies on the \code{OpenStreetMap} package to obtain base
layer tiles. This in turn requires Java to be installed, and linking R with
Java can sometimes prove challenging. On macOS, Java errors can often be
resolved by entering \verb{sudo R CMD javareconf} in a terminal ( which updates
R's Java configuration variables). On Windows, ensure that the Java
architecture (32-bit/64-bit) matches that of R. Additionally, some Java
errors arise when using RStudio but not when using R.
}
