% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotic_weight.R
\name{suitability}
\alias{suitability}
\title{Combine abiotic and biotic layers to create suitability raster}
\usage{
suitability(x, outfile, return_rast = FALSE)
}
\arguments{
\item{x}{One of: a \code{\link{SpatRaster}} object; a \code{RasterLayer}, a list of
\code{\link{SpatRaster}} objects, or a vector of file paths to raster files.}

\item{outfile}{Character. Filename where output will be saved. If not
provided, \code{\link{SpatRaster}} object will be returned to R.}

\item{return_rast}{Logical. Should the \code{\link{SpatRaster}} object be returned to
R? Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting \code{\link{SpatRaster}} object is
saved as to that path. If \code{return_rast} is \code{TRUE} or \code{outfile} is not
specified the resulting \code{\link{SpatRaster}} is returned, otherwise \code{outfile} is
returned invisibly.
}
\description{
Combine biotic (and abiotic) layers into a single suitability raster to be
used to scale arrival rates.
}
