% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_coverage.R
\name{surveillance_coverage}
\alias{surveillance_coverage}
\title{Calculate the proportion of total establishment likelihood captured by
current surveillance}
\usage{
surveillance_coverage(ras, layer_names, surveillance_locations)
}
\arguments{
\item{ras}{\verb{Raster*}, \code{\link{SpatRaster}}, or a character vector of one or more
paths to raster files describing establishment likelihood.}

\item{layer_names}{Character. Name to be assigned to each establishment
raster. If not specified, names of the raster layers will be used.}

\item{surveillance_locations}{Character. Path to a csv file containing
surveillance locations, including columns titled "Longitude" and
"Latitude" with coordinates given in decimal degrees (GDA94).}
}
\value{
A \code{data.frame} with the proportion of establishment likelihood
captured by current surveillance.
}
\description{
Calculate the proportion of total establishment likelihood captured by
current surveillance.
}
