% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals_by_tourists.R
\name{arrivals_by_tourists}
\alias{arrivals_by_tourists}
\title{Estimates pest arrival due to tourists}
\usage{
arrivals_by_tourists(
  tourist_beds,
  airport_weights,
  total_tourists,
  probability,
  outfile,
  return_rast = FALSE
)
}
\arguments{
\item{tourist_beds}{A \code{RasterLayer} or file path to a raster file 
containing tourist bed numbers.}

\item{airport_weights}{A \code{RasterLayer} or file path to a raster file 
containing airport distance weights.}

\item{total_tourists}{Integer. The number of tourists.}

\item{probability}{Numeric. The probability a tourist carries a pest.}

\item{outfile}{Character. Output raster file path. If not provided, raster 
object will be returned to R.}

\item{return_rast}{Logical. Should the raster object be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting raster is saved as a
  to that path. If \code{return_rast} is \code{TRUE} or \code{outfile} is 
  not specified the resulting raster is returned, otherwise \code{NULL} is 
  returned invisibly.
}
\description{
Estimates pest arrival due to tourists as a function of distance from airport
and tourist accommodation.
}
\seealso{
Other functions estimating arrivals: 
\code{\link{arrivals_by_containers}()},
\code{\link{arrivals_by_fertiliser}()},
\code{\link{arrivals_by_goods}()},
\code{\link{arrivals_by_machinery}()},
\code{\link{arrivals_by_mail}()},
\code{\link{arrivals_by_nurserystock}()},
\code{\link{arrivals_by_residents}()},
\code{\link{arrivals_by_torres}()},
\code{\link{arrivals_by_vessels}()}
}
\concept{functions estimating arrivals}
