% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_median.R
\name{calculate_median}
\alias{calculate_median}
\title{Calculate median of raster objects}
\usage{
calculate_median(x, outfile)
}
\arguments{
\item{x}{Either a \code{RasterStack}, a multi-layer  \code{\link{SpatRaster}}, or a vector
of file paths to raster files.}

\item{outfile}{Character. Optional output raster file path. Directory will be
created recursively if it does not exist.}
}
\value{
If \code{outfile} is specified, the resulting \code{\link{SpatRaster}} is saved to
that path and \code{outfile} is returned invisibly. If \code{outfile} is not
specified the \code{\link{SpatRaster}} is returned to R.
}
\description{
Calculate the cellwise median of multiple rasters.
}
