% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captured_by_ncells.R
\name{captured_by_ncells}
\alias{captured_by_ncells}
\title{Calculate establishment likelihood captured in top n cells}
\usage{
captured_by_ncells(x, layer_names, n_cells, all = TRUE)
}
\arguments{
\item{x}{\verb{Raster*} or \code{SpatRaster} object, or a character vector of file
path(s) to one or more raster files.}

\item{layer_names}{Character vector. Optional names corresponding to \code{x}.}

\item{n_cells}{Integer. The number of cells to consider.}

\item{all}{Logical. If \code{TRUE}, return the proportion captured from 1 to \emph{n}
cells. If \code{FALSE}, return the cumulative proportion.}
}
\value{
Proportion of establishment likelihood captured, or a vector of
cumulative proportions.
}
\description{
Calculate the proportion of establishment likelihood captured in top n cells.
}
