% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establishment_likelihood.R
\name{establishment_likelihood}
\alias{establishment_likelihood}
\title{Estimate establishment likelihood}
\usage{
establishment_likelihood(
  total_arrivals,
  suitability,
  outfile,
  return_rast = FALSE
)
}
\arguments{
\item{total_arrivals}{A \code{RasterLayer} (or path to raster file)
containing total arrival estimates.}

\item{suitability}{A \code{RasterLayer} (or path to raster file) containing
suitability scores.}

\item{outfile}{Character. Output raster file path. If not provided, 
the \code{RasterLayer} will be returned to R.}

\item{return_rast}{Logical. Should the \code{RasterLayer} be returned to R?
Ignored if \code{outfile} is not provided.}
}
\value{
If \code{outfile} is specified, the resulting \code{RasterLayer} is 
  saved as to that path. If \code{return_rast} is \code{TRUE} or
  \code{outfile} is not specified the resulting \code{RasterLayer} is 
  returned, otherwise \code{NULL} is returned invisibly.
}
\description{
Estimate the likelihood of pest establishment based on total arrivals and
environmental suitability.
}
