% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_worldclim2.R
\name{extract_worldclim2}
\alias{extract_worldclim2}
\title{Extract WorldClim 2.0 data}
\usage{
extract_worldclim2(path_2_zip, outdir)
}
\arguments{
\item{path_2_zip}{Character. Path to .zip archive downloaded by 
\code{\link{download_worldclim2}}.}

\item{outdir}{Character. File path to which contained files should be 
extracted. Will be created (recursively) if necessary.}
}
\value{
Raster data are extracted to \code{outdir} and \code{NULL} is
  returned invisibly.
}
\description{
Extract WorldClim 2.0 data
}
\examples{
\dontrun{
download_worldclim2('bioclim_10m.zip', 'bio', '10m')
extract_worldclim2('bioclim_10m.zip', outdir='bioclim')
}
}
\seealso{
\code{\link{download_worldclim2}}
}
