% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_records.R
\name{get_gbif_records}
\alias{get_gbif_records}
\title{Downloads GBIF records of species}
\usage{
get_gbif_records(species, limit = 2e+05, min_year, coord_uncertainty)
}
\arguments{
\item{species}{Character. Species taxonomic name.}

\item{limit}{Integer. Number of records to return. Max \code{limit} is
\code{200000}.}

\item{min_year}{Integer. The minimum year for which records should be
collated. Default is \code{NULL}, i.e. no minimum.}

\item{coord_uncertainty}{Integer. The maximum allowable documented 
coordinate uncertainty (in metres). If specified, this argument will 
result in removal of any records that have a documented uncertainty above 
this limit. Records with no uncertainty documented will be returned 
regardless of the value of \code{coord_uncertainty}.}
}
\value{
A \code{data.frame} of species occurrence records.
}
\description{
Downloads GBIF records of species, removing auxiliary columns and records 
recorded prior to a specified year, or that have coordinate uncertainty
above a specified amount.
}
\details{
This function is a wrapper of \code{rgbif} such that it can be
  readily used with \code{CoordinateCleaner} package.
}
