% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_records.R
\name{get_gbif_records}
\alias{get_gbif_records}
\title{Downloads GBIF records of species}
\usage{
get_gbif_records(
  taxon,
  min_year,
  coord_uncertainty,
  method = c("search", "download"),
  username,
  pwd,
  email
)
}
\arguments{
\item{taxon}{Character. Species taxonomic name.}

\item{min_year}{Integer. The minimum year for which records should be
collated. Default is `NULL`, i.e. no minimum.}

\item{coord_uncertainty}{Integer. The maximum allowable documented coordinate
uncertainty (in metres). If specified, this argument will result in removal
of any records that have a documented uncertainty above this limit. Records
with no uncertainty documented will be returned regardless of the value of
`coord_uncertainty`.}

\item{method}{Either `'search'` (uses the GBIF `/occurrence/search` API
endpoint) or `'download'` (uses the GBIF `/occurrence/download`API
endpoint). The former makes paginated queries to the API, while the latter
performs an asynchronous query (but waits for the resulting dataset to be
ready for download). The `'search'` method is limited to 100,000 records;
for large datasets, consider using `'download'`. When using 
`method='download'`, the arguments `username`, `pwd`, and `email` must be 
provided.}

\item{username}{An email address for GBIF notification when download is 
ready, required when method is `'download'`.}

\item{pwd}{GBIF password, required when method is `'download'`.}

\item{email}{Email address, required when `method = 'download'`. This _may_ 
be used to notify user when download is ready.}
}
\value{
A `data.frame` of species occurrence records.
}
\description{
Downloads GBIF records of species, removing auxiliary columns and records 
recorded prior to a specified year, or that have coordinate uncertainty
above a specified amount.
}
\details{
This function is a wrapper of `rgbif` such that it can be readily
  used with the `CoordinateCleaner` package.
}
