% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncells_4_threshold.R
\name{ncells_4_threshold}
\alias{ncells_4_threshold}
\title{Extract number of cells required to meet threshold proportion of risk}
\usage{
ncells_4_threshold(
  risk_rasters,
  layer_names,
  proportion_captured = c(0.6, 0.8, 0.9, 0.95)
)
}
\arguments{
\item{risk_rasters}{\verb{Raster*}, \code{\link{SpatRaster}}, or character vector of one or
more pathers to raster files.}

\item{proportion_captured}{Numeric vector. Proportion(s) of risk to be
captured.}

\item{names}{Character. Optional names corresponding to \code{risk_rasters}. If
not provided, layer names are used.}
}
\value{
A \code{tibble} containing the number of cells requiring surveillance in
order to capture the proportion(s) of total risk given by
\code{proportion_captured}.
}
\description{
Extract number of cells required to meet threshold proportion of risk.
}
