% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_raster}
\alias{plot_raster}
\title{Plot a raster}
\usage{
plot_raster(
  object,
  legend_title,
  occurrence_data = NULL,
  pt_col = "red",
  height,
  units = c("in", "cm", "mm"),
  compass = FALSE,
  outfile
)
}
\arguments{
\item{object}{A \code{RasterLayer} or file path to raster file.}

\item{legend_title}{Character. If missing, the name of the raster layer will
be used.}

\item{occurrence_data}{A \code{data.frame}, \code{sf} object,
\code{SpatialPointsDataFrame} object, or path to a .csv file containing
columns named "Latitude" and "Longitude". If \code{NULL}, no points will 
be plotted.}

\item{pt_col}{Character. Colour of points (if plotted).}

\item{height}{Height of plot, specified in \code{units}. Required if
\code{outfile} is provided.}

\item{units}{Character. Units corresponding to \code{height} and 
\code{width}. Can be \code{"in"}, \code{"cm"}, or \code{"mm"}. Default is 
inches (\code{"in"}).}

\item{compass}{Logical. Should a North arrow be shown?}

\item{outfile}{Character. Path to save output.}
}
\value{
A \code{tmap} object. If \code{outfile} is provided, a map will also
  be written to that file.
}
\description{
Plot a raster.
}
