% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_compress.R
\name{rle_compress}
\alias{rle_compress}
\title{Compress raster data using run length encoding}
\usage{
rle_compress(x, outfile, quiet = FALSE, overwrite = FALSE)
}
\arguments{
\item{x}{Either a \code{RasterLayer} object, single-layer \code{\link{SpatRaster}} object,
or file path to a categorical raster to be compressed.}

\item{outfile}{Character (optional). Path to target .rds file that will
store run-length encoded results. Directory will be created recursively if
it doesn't exist.}

\item{quiet}{Logical. Should progress messages be suppressed?}

\item{overwrite}{Logical. Should \code{outfile} be overwritten if it already
exists? Ignored if \code{outfile} is not provided.}
}
\value{
A \code{raster_rle} object (extending \code{tibble}) with three columns
(\code{length}, \code{value}, and \code{start}) and attributes (\code{extent}, a
\code{\link{SpatExtent}} object, \code{res}, and \code{crs}). This object is additionally saved
in rds format to \code{outfile}, if provided.
}
\description{
Compress categorical raster data using run length encoding.
}
