% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_map.R
\name{static_map}
\alias{static_map}
\title{Produce a static map of establishment likelihood}
\usage{
static_map(ras, xlim, ylim, layer, legend_title, set_value_range,
  scale_type = "none", transparency = 0.7,
  colramp_entire_range = TRUE, surveillance_locs, pt_col = "red",
  aggregate_raster, height, units = c("in", "cm", "mm"), outfile)
}
\arguments{
\item{ras}{A \code{RasterLayer} or file path to a raster file.}

\item{xlim}{Numeric vector. The longitudinal extent of the area to plot.}

\item{ylim}{Numeric vector. The latitudinal extent of the area to plot.}

\item{layer}{Character. A layer name to be plotted. Only relevant if loading
a \code{RasterStack} or \code{RasterBrick}.}

\item{legend_title}{Character. Legend title.}

\item{set_value_range}{A numeric vector containing an upper and lower bound
value (in units of raster). Values outside this range (including values
falling on the boundaries) will be masked.}

\item{scale_type}{Character. Can be: \code{"none"} (raw data, no rescaling),
\code{"log10"}, \code{"max normalize"} (proportional to maximum value),
\code{"minmax normalize"} (rescale values to be between zero and 1 based 
on min and max), or \code{"logit"}. Note that if \code{"log10"} or 
\code{"logit"} is used 0 or 1 values. must be masked (using 
\code{set_value_range}) or rescaled outside of this function.}

\item{transparency}{Numeric. Transparency of raster, between 0-1.}

\item{colramp_entire_range}{Logical. Whether to set colour ramp limits based
on national risk range (\code{TRUE}) or by risk range present in region
specified by \code{xlim} and \code{ylim}.}

\item{surveillance_locs}{A spatial object or a path to a .csv file 
containing columns named "Latitude" and "Longitude".}

\item{pt_col}{Character. The plotting colour for surveillance points.}

\item{aggregate_raster}{\code{NULL} or a list containing the aggregation
factor (i.e. number of raster cells to aggregate) and the aggregation
function e.g. \code{list(10, sum)}.}

\item{height}{height of plot. If not defined will use size of current 
graphic device. Width will be dertermined automatically, based on the 
aspect ratio given by the plotting extent.}

\item{units}{Character. Units in which \code{height} is specified. Can be 
\code{"in"} (inches), \code{"cm"}, or \code{"mm"}. Default is \code{"in"}.}

\item{outfile}{Character. File path to save map.}
}
\value{
If \code{outfile} is provided, a map is saved to that file. 
  Otherwise, a \code{tmap} object is returned to R.
}
\description{
Produce a static map of establishment likelihood, with OpenStreetMap base 
layer.
}
\details{
This function relies on the \code{OpenStreetMap} package to obtain 
  base layer tiles. This in turn requires Java to be installed, and linking 
  R with Java can sometimes prove challenging. On macOS, Java errors can 
  often be resolved by entering \code{sudo R CMD javareconf} in a terminal (
  which updates R's Java configuration variables). On Windows, ensure that 
  the Java architecture (32-bit/64-bit) matches that of R. Additionally, 
  some Java errors arise when using RStudio but not when using R.
}
