% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_function.R
\name{step_function}
\alias{step_function}
\title{Apply a step function to raster values}
\usage{
step_function(rast, threshold, lower_value, upper_value)
}
\arguments{
\item{rast}{A \verb{Raster*} or \code{\link{SpatRaster}} object.}

\item{threshold}{Numeric. The threshold value.}

\item{lower_value}{Numeric. The value to use below the threshold.}

\item{upper_value}{Numeric. The value to use above the threshold.}
}
\value{
A \code{\link{SpatRaster}} object.
}
\description{
Reclassify raster values either side of a defined threshold value.
}
