% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_coverage.R
\name{surveillance_coverage}
\alias{surveillance_coverage}
\title{Calculate the proportion of total establishment likelihood captured by
current surveillance}
\usage{
surveillance_coverage(
  establishment_rasters,
  layer_names,
  surveillance_locations
)
}
\arguments{
\item{establishment_rasters}{Character vector. A vector of one or more file 
path to rasters describing establishment likelihood.}

\item{layer_names}{Character. Name to be assigned to each establishment
raster. If not specified, names of the raster layers will be used.}

\item{surveillance_locations}{Character. Path to a csv file containing 
surveillance locations, including columns titled "Longitude" and 
"Latitude" with coordinates given in decimal degrees (GDA94).}
}
\value{
The proportion of establishment likelihood captured by current
  surveillance.
}
\description{
Calculate the proportion of total establishment likelihood captured by 
current surveillance.
}
